backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
pydantic[email]==2.7.1
aiosqlite==0.20.0
numpy==1.26.4
python-multipart==0.0.9
torch==2.2.0+cpu
